requirements.txt
langchain_openai
langchain
langchain_community
python-dotenv
langchain-chroma